backend-security-auditor/requirements.txt
gitpython>=3.1.0
pydantic>=2.0.0
